+++ hub/__tests__/escapeHtml.test.js
@@ -1,0 +1,6 @@
+import { escapeHtml } from "../src/utils.js";
+
+test("escapeHtml escapes &, < and >", () => {
+  expect(escapeHtml("&<>")).toBe("&amp;&lt;&gt;");
+});
+